requirements.txt
pyyaml
jinja2
requests
yfinance
pandas
websockets
pydantic
pydantic-settings
tenacity
lxml
